Refactor stories fodler (#450)
