+++ dashboard/src/components/common/Header/Header.tsx
@@ -1,4 +1,4 @@
-import { Button } from "./Button";
+import { Button } from "../Button/Button";
 import "./header.css";
 
 type User = {
