+++ dashboard/src/stories/Introduction.stories.mdx
@@ -1,12 +1,12 @@
 import { Meta } from "@storybook/addon-docs";
-import Code from "./assets/code-brackets.svg";
-import Colors from "./assets/colors.svg";
-import Comments from "./assets/comments.svg";
-import Direction from "./assets/direction.svg";
-import Flow from "./assets/flow.svg";
-import Plugin from "./assets/plugin.svg";
-import Repo from "./assets/repo.svg";
-import StackAlt from "./assets/stackalt.svg";
+import Code from "../assets/code-brackets.svg";
+import Colors from "../assets/colors.svg";
+import Comments from "../assets/comments.svg";
+import Direction from "../assets/direction.svg";
+import Flow from "../assets/flow.svg";
+import Plugin from "../assets/plugin.svg";
+import Repo from "../assets/repo.svg";
+import StackAlt from "../assets/stackalt.svg";
 
 <Meta title="Example/Introduction" />
